restructure code layout
